requirements.txt
fastapi==0.135.2
imbalanced_learn==0.11.0
imbalanced_learn==0.14.1
joblib==1.2.0
joblib==1.5.3
network==0.1
numpy==2.4.3
paho_mqtt==2.1.0
pandas==3.0.1
Requests==2.33.0
scikit_learn==1.2.2
streamlit==1.46.0
streamlit==1.30.0
streamlit==1.55.0
xgboost==3.2.0
